streamlit_ui/requirements.txt
streamlit
google-cloud-dialogflow-cx
google-auth
google-cloud-core
google-cloud
